backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.35.0
httpx==0.28.1
python-multipart==0.0.20
pydantic-settings==2.10.1
pillow==11.3.0
pytesseract==0.3.13
rapidocr-onnxruntime==1.4.4
